twitter-to-pubsub/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-pubsub==2.13.6
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
oauthlib==3.2.0
proto-plus==1.22.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tweepy==4.10.0
urllib3==1.26.11
